refactor(di): introduce PluginBootstrapper as composition root
Create a dedicated bootstrapper class that centralises all DDD service
hook wiring, following the composition root pattern. This reduces
scattered Container::instance() calls throughout the plugin and provides
a single location for understanding how services connect to WordPress.

The bootstrapper handles:
- Shortcode filter registration
- Content filter system (InputSanitizer, ContentFilterRegistry)
- Key events (configuration, service, shortcode handler)
- Lazyload configuration
- Auto-archive cron handler

Also replace ContainerAwareTrait with IntegrationTestCase base class,
providing a cleaner inheritance-based approach for test container access.
