+++ liveblog.php
@@ -273,8 +273,6 @@ if ( ! class_exists( 'WPCOM_Liveblog' ) ) :
 			self::register_embed_handlers();
 
 			WPCOM_Liveblog_Entry_Key_Events_Widget::load();
-			self::init_key_events();
-			self::init_lazyload();
 			WPCOM_Liveblog_Socketio_Loader::load();
 			WPCOM_Liveblog_Entry_Embed_SDKs::load();
 			WPCOM_Liveblog_AMP::load();
@@ -283,8 +281,10 @@ if ( ! class_exists( 'WPCOM_Liveblog' ) ) :
 				WPCOM_Liveblog_Rest_Api::load();
 			}
 
-			// Activate the WP CRON Hooks.
-			\Automattic\Liveblog\Infrastructure\DI\Container::instance()->auto_archive_cron_handler()->register();
+			// Initialize DDD services via the bootstrapper (centralises hook wiring).
+			$container    = \Automattic\Liveblog\Infrastructure\DI\Container::instance();
+			$bootstrapper = new \Automattic\Liveblog\Infrastructure\WordPress\PluginBootstrapper( $container );
+			$bootstrapper->init();
 		}
 
 		/**
@@ -388,180 +388,9 @@ if ( ! class_exists( 'WPCOM_Liveblog' ) ) :
 			add_filter( 'comment_class', array( __CLASS__, 'add_comment_class' ), 10, 3 );
 			add_filter( 'is_protected_meta', array( __CLASS__, 'protect_liveblog_meta_key' ), 10, 2 );
 
-			// Add In the Filter hooks to Strip any Restricted Shortcodes before a new post or updating a post.
-			$shortcode_filter = \Automattic\Liveblog\Infrastructure\DI\Container::instance()->shortcode_filter();
-			add_filter( 'liveblog_before_insert_entry', array( $shortcode_filter, 'filter' ), 10, 1 );
-			add_filter( 'liveblog_before_update_entry', array( $shortcode_filter, 'filter' ), 10, 1 );
-
 			// We need to check the Liveblog autoarchive date on each time a new entry is added or updated to
 			// ensure we extend the date out correctly to the next archive point based on the configured offset.
 			add_filter( 'liveblog_before_insert_entry', array( __CLASS__, 'update_autoarchive_expiry' ), 10, 1 );
-
-			// Initialise DDD content filters (InputSanitizer and ContentFilterRegistry).
-			self::init_content_filters();
-		}
-
-		/**
-		 * Initialise the DDD-based content filter system.
-		 *
-		 * Registers all content filters with the ContentFilterRegistry and sets up
-		 * the input sanitizer hooks. Handles commands, emojis, hashtags, and authors.
-		 */
-		private static function init_content_filters(): void {
-			$container = \Automattic\Liveblog\Infrastructure\DI\Container::instance();
-
-			// Register all content filters with the registry.
-			$registry = $container->content_filter_registry();
-			$registry->register( $container->command_filter() );
-			$registry->register( $container->emoji_filter() );
-			$registry->register( $container->hashtag_filter() );
-			$registry->register( $container->author_filter() );
-
-			// Initialise the registry (sets up prefixes, regex, loads filters).
-			$registry->initialise( 'commands, emojis, hashtags, authors' );
-
-			// Add input sanitizer hooks (runs before content filters).
-			$input_sanitizer = $container->input_sanitizer();
-			add_filter( 'liveblog_before_insert_entry', array( $input_sanitizer, 'sanitize' ), 1 );
-			add_filter( 'liveblog_before_update_entry', array( $input_sanitizer, 'sanitize' ), 1 );
-			add_filter( 'liveblog_before_insert_entry', array( $input_sanitizer, 'fix_links_wrapped_in_div' ), 1 );
-			add_filter( 'liveblog_before_update_entry', array( $input_sanitizer, 'fix_links_wrapped_in_div' ), 1 );
-			add_filter( 'liveblog_before_preview_entry', array( $input_sanitizer, 'fix_links_wrapped_in_div' ), 1 );
-
-			// Add content filter hooks (processes commands, emojis, hashtags, authors).
-			add_filter( 'liveblog_before_insert_entry', array( $registry, 'apply_filters' ), 10 );
-			add_filter( 'liveblog_before_update_entry', array( $registry, 'apply_filters' ), 10 );
-			add_filter( 'liveblog_before_preview_entry', array( $registry, 'apply_filters' ), 10 );
-
-			// Add revert filter hook (for edit mode).
-			add_filter( 'liveblog_before_edit_entry', array( $registry, 'revert_all' ), 10 );
-
-			// Allow emoji image attributes (class, width, height, data-emoji) to pass through.
-			add_filter(
-				'liveblog_image_allowed_attributes',
-				function ( $attrs ) {
-					return array_merge( $attrs, array( 'class', 'width', 'height', 'data-emoji' ) );
-				}
-			);
-		}
-
-		/**
-		 * Initialise the DDD-based key events system.
-		 *
-		 * Sets up the key event configuration, shortcode, and all related hooks.
-		 */
-		private static function init_key_events(): void {
-			$container = \Automattic\Liveblog\Infrastructure\DI\Container::instance();
-
-			$configuration     = $container->key_event_configuration();
-			$key_event_service = $container->key_event_service();
-			$shortcode_handler = $container->key_event_shortcode_handler();
-
-			// Initialize key event configuration (sets up templates/formats).
-			add_action( 'init', array( $configuration, 'initialize' ), 11 );
-
-			// Register the /key command.
-			add_filter(
-				'liveblog_active_commands',
-				function ( $commands ) {
-					$commands[] = 'key';
-					return $commands;
-				}
-			);
-
-			// Enrich entry JSON with key event data.
-			add_filter(
-				'liveblog_entry_for_json',
-				function ( $entry, $entry_object ) use ( $key_event_service, $configuration ) {
-					// Support both legacy WPCOM_Liveblog_Entry and domain Entry objects.
-					if ( $entry_object instanceof \Automattic\Liveblog\Domain\Entity\Entry ) {
-						$post_id = $entry_object->post_id();
-						return $key_event_service->enrich_entry_for_json( $entry, $entry_object, $post_id );
-					}
-
-					// Handle legacy WPCOM_Liveblog_Entry.
-					$post_id = $entry_object->get_post_id();
-					$content = $entry_object->get_content();
-
-					// Detect key event from content (more reliable than meta for updates).
-					$is_key_event       = $key_event_service->content_has_key_command( $content );
-					$entry['key_event'] = $is_key_event;
-
-					if ( $is_key_event ) {
-						$entry['key_event_content'] = $configuration->format_content( $content, $post_id );
-					}
-
-					return $entry;
-				},
-				10,
-				2
-			);
-
-			// Handle /key command action (add meta when command is used).
-			add_action(
-				'liveblog_command_key_after',
-				function ( $content, $entry_id, $post_id ) use ( $key_event_service ) {
-					$key_event_service->handle_key_command( $content, $entry_id, $post_id );
-				},
-				10,
-				3
-			);
-
-			// Sync key event meta when entry is updated.
-			add_action(
-				'liveblog_update_entry',
-				array( $key_event_service, 'sync_key_event_meta' ),
-				10,
-				2
-			);
-
-			// Add admin options for key event settings.
-			add_filter(
-				'liveblog_admin_add_settings',
-				function ( $extra_fields, $post_id ) use ( $shortcode_handler ) {
-					$extra_fields[] = $shortcode_handler->get_admin_options( $post_id );
-					return $extra_fields;
-				},
-				10,
-				2
-			);
-
-			// Save admin options.
-			add_action(
-				'liveblog_admin_settings_update',
-				function ( $response, $post_id ) use ( $shortcode_handler ) {
-					$shortcode_handler->save_admin_options( $response, $post_id );
-				},
-				10,
-				2
-			);
-
-			// Register the shortcode.
-			$shortcode_handler->register();
-
-			// Hook configuration for content formatting.
-			add_filter(
-				'liveblog_key_event_content',
-				function ( $content, $post_id ) use ( $configuration ) {
-					return $configuration->format_content( $content, $post_id );
-				},
-				10,
-				2
-			);
-		}
-
-		/**
-		 * Initialise the DDD-based lazyload configuration.
-		 *
-		 * Sets up lazy loading configuration for entries including initial
-		 * display count and per-page limits for lazy loading requests.
-		 */
-		private static function init_lazyload(): void {
-			$container     = \Automattic\Liveblog\Infrastructure\DI\Container::instance();
-			$configuration = $container->lazyload_configuration();
-
-			// Initialize on template_redirect when liveblog state is available.
-			add_action( 'template_redirect', array( $configuration, 'initialize' ) );
 		}
 
 		/**
